simple-deep-research/requirements.txt
openai-agents
pydantic
python-dotenv